passwords/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
zxcvbn>=4.4.28
argon2-cffi>=23.1.0
